Add partial functionality for actuator states
Added functionality for the fire valve and the quick disconnect. Not sure how to actually test if they work. XV valves might need some extra functionality, as the names should change based on which valves are currently being used. That's my interpretation, at least. If that is not the case, code them in as is.