Actuator states (#18)
* Add parts in ui for actuator states

* Add partial functionality for actuator states

Added functionality for the fire valve and the quick disconnect. Not sure how to actually test if they work. XV valves might need some extra functionality, as the names should change based on which valves are currently being used. That's my interpretation, at least. If that is not the case, code them in as is.

* Added more valves

Increased number of XV valves from 4 to 12

* Made log output bigger

* Commented out stylesheets for later use

* Changed if statement for actuator state header

* More comments

* Fix control layout spacing

* Add tabbing to UI

---------

Co-authored-by: VorteX254 <135150065+VorteX254@users.noreply.github.com>
Co-authored-by: Elias Hawa <elias.hawa2468@gmail.com>