Pid diagram #9 (#30)
* Add parts in ui for actuator states

* Add partial functionality for actuator states

Added functionality for the fire valve and the quick disconnect. Not sure how to actually test if they work. XV valves might need some extra functionality, as the names should change based on which valves are currently being used. That's my interpretation, at least. If that is not the case, code them in as is.

* Added more valves

Increased number of XV valves from 4 to 12

* Made log output bigger

* Commented out stylesheets for later use

* Changed if statement for actuator state header

* More comments

* Rebase completed

* Add parts in ui for actuator states
Rebase done!

* Added more valves

Increased number of XV valves from 4 to 12

* Made log output bigger

* Commented out stylesheets for later use

* Changed if statement for actuator state header

* Rebase completed

* Added PID diagram and labels

* Make P&I diagram a resource file

* Update form.ui

* Deleted duplicate UDP connection button at the bottom of ui

* Remove comments

---------

Co-authored-by: jackricketts <jackricketts@rogers.com>
Co-authored-by: Elias Hawa <elias.hawa2468@gmail.com>